update bar chart for debt

+++ app/models/asset.js
@@ -2,9 +2,9 @@ module.exports = function (sequelize, DataTypes) {
 
   var Asset = sequelize.define('Asset', {
     relation      : DataTypes.STRING,       // 본인과의 관계
-    change        : DataTypes.INTEGER,        // 증감액
-    prevTotal     : DataTypes.INTEGER,  // 종전가액
-    total         : DataTypes.INTEGER,         // 현재가액
+    change        : DataTypes.BIGINT,        // 증감액
+    prevTotal     : DataTypes.BIGINT,  // 종전가액
+    total         : DataTypes.BIGINT,         // 현재가액
     description   : DataTypes.TEXT,    // 소재지 면적 등 권리 등의 명세
     reason        : DataTypes.TEXT            // 변동 사유
   }, {
@@ -18,4 +18,3 @@ module.exports = function (sequelize, DataTypes) {
 
   return Asset
 }
-
